+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 - Quick fixes for null check and equality tests for _UnusedAssertkAssertion_
-- Quick fixes for common junit 4 assertion cases detected by _TestFrameworkAssertionDetector_
+- Quick fixes for common junit 4 assertions detected by _TestFrameworkAssertionDetector_
+- Quick fixes for common kotlin.test assertions cases detected by _TestFrameworkAssertionDetector_
 
 ### Changed